requirements.txt
fastapi
uvicorn
python-dotenv
jira
requests
pydantic
